Fix type

+++ src/main/python/emr_autoscaling/emr.py
@@ -84,7 +84,7 @@ class Emr:
             reverse = True
         )
         for group in groups:
-            target_requested_instances = group["RequestedInstanceCount"] + math.ceil(direction * 0.2 * group["RequestedInstanceCount"])
+            target_requested_instances = int(group["RequestedInstanceCount"] + math.ceil(direction * 0.2 * group["RequestedInstanceCount"]))
             if self.min_instances <= target_requested_instances <= self.max_instances:
                 self.emr.modify_instance_groups (
                     InstanceGroups = [
